Updated code from jeremy boynes...
========================================================================
I was doing some background hacking on the infomodel classes from Scout
to get a bit closer to the spec, to implement some of the collections
stuff, and to get rid of what I assumed were legacy references to
uddi4j. It turned into a bigger change than I expected and I didn't see
a lot of activity on Scout so I was wondering what to do with it.
========================================================================


git-svn-id: https://svn.apache.org/repos/asf/webservices/scout/trunk@57093 13f79535-47bb-0310-9956-ffa450edef68