SCOUT-128, set the UDDI version to "3.0" if 'v3' is found in the connection URL.
In some tests the TCK connects directly to scout, not using the JNDI lookup. In this case the UDDI version property is not set, and Scout defaults it to "2.0", which leads to issues. The preferred way to solve this is to have the TCK set this property, but this does seem that easy to do. To work around this issue we check the juddi URL, and if this contains 'v3' we automatically set the UDDI version to "3.0" if it is not set. This maybe a jUDDI specific fix. I'm open to other suggestions.

git-svn-id: https://svn.apache.org/repos/asf/juddi/scout/trunk@1295121 13f79535-47bb-0310-9956-ffa450edef68